Adjusted jetpack tiers
The main thing I did here was a remove a tier so there are now 3 jetpack tiers, and the recipes are adjusted based on that.

Tier 1 is a combination of the old tiers 1 and 2 recipes, while tier 2 is essentially the old tier 3 recipe. Each of them has the same thrust as before, so they're basically just more expensive than they were previously.
Tier 3 is the old tier 4, even including the thrust value and a similar overall price. I also made the cube cost of this more expensive (it was actually like this in the first commit I did. I toned it down in the "various fixes" commit, but decided to revert it since I feel the composite is a more fitting ingredient next to the modules)